requirements.txt
avro-python3==1.9.1
certifi==2019.11.28
chardet==3.0.4
confluent-kafka[avro]==1.3.0
fastavro==0.22.9
idna==2.8
protobuf==3.11.2
requests==2.22.0
six==1.14.0
urllib3==1.25.8
